requirements.txt
# web framework and API server
fastapi==0.110.0  
uvicorn[standard]==0.29.0   # ASGI server for running FastAPI applications
pydantic==2.6.4  # Data validation and settings management

# HTTP client and API communication
httpx==0.27.0  # Async HTTP client for making API calls
tenacity==8.2.3  # Retry library for handling transient failures

# vector database and embeddings
chromadb==0.4.24 # Stores document embeddings and enables fast similarity search
sentence-transformers==2.7.0   # Converts text documents to numerical vectors for similarity search

# environment and configuration 
python-dotenv==1.0.1   # Loads environment variables from .env files for configuration
orjson==3.10.1  #Improves performance of JSON operations in our API

# optional enhancment 
redis==5.0.3   # Caching for improved performance and rate limiting
numpy==1.26.4 